Create NetConfig abstract class
